LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/update-recovery.https-expected.txt

CONSOLE MESSAGE: FetchEvent.respondWith received an error: undefined
CONSOLE MESSAGE: Cannot load https://web-platform.test:9443/service-workers/service-worker/resources/simple.txt.

PASS Recover from a bad service worker by updating after a failed navigation.
